// src/app/models/api.ts
export interface ApiResponse<T> {
  success: boolean;
  data?: T | null;
  message?: string | null;
  errors?: string[] | null;
}
// ...
export interface CashRegisterDto {
  id?: number;
  name: string;
  isOpen: boolean;
  balance: number;
  companyId?: number;
}

// CashMovement
export interface CashMovementDto {
  id?: number;
  cashRegisterId: number;
  amount: number;
  type: string;
  description?: string | null;
  creationDate?: string;
}